feat(getData): add kultCast integration to mark cast members
This commit introduces the `kultCast` array to the `getFilmCastData` function, enabling the identification of cast members who are part of the `kultCast` list. If a cast member is found in `kultCast`, a `kult` property is added and set to `true` for that member. This change enhances the functionality by providing additional context about cast members.